refactor architecture, font and menu toggle
