+++ components/layout/layout.tsx
@@ -3,12 +3,12 @@ import Header from "./header"
 import Footer from "./footer"
 import Sidebar from "./sidebar"
 import { Oxygen } from "next/font/google"
-import styles from "@/styles/Layout.module.scss"
+import styles from "@/styles/layout/Layout.module.scss"
 import localFont from "next/font/local"
 import ChatButton from "./chatButton"
 
 const bigJohnPro = localFont({
-  src: "../public/fonts/bjp-bold.otf",
+  src: "../../public/fonts/bjp-bold.otf",
   variable: "--bigJohnPro-font",
   display: "swap",
   fallback: ["system-ui", "arial", "sans-serif"],
@@ -25,13 +25,26 @@ const oxygen = Oxygen({
 export default function Layout({ children }: PropsWithChildren) {
   const [activeMenu, setActiveMenu] = useState<boolean>(false)
 
+  const handleActiveMenu = () => {
+    setActiveMenu(!activeMenu)
+  }
+
   return (
     <>
-      <Header setActiveMenu={setActiveMenu} />
-      <Sidebar setActiveMenu={setActiveMenu} activeMenu={activeMenu} />
+      <Header 
+        onClick={handleActiveMenu}
+        font={bigJohnPro} 
+      />
+      <Sidebar 
+        onClick={handleActiveMenu}
+        activeMenu={activeMenu} 
+        font={bigJohnPro}
+      />
       <main className={`${styles.main} ${oxygen.className} ${bigJohnPro.variable}`}>{children}</main>
       <ChatButton />
-      <Footer />
+      <Footer 
+        font={oxygen}
+      />
     </>
   )
 }
