+++ pages/_app.tsx
@@ -2,7 +2,7 @@ import { appWithTranslation } from "next-i18next"
 import "@/styles/globals.scss"
 import type { AppProps } from "next/app"
 import { Analytics } from "@vercel/analytics/react"
-import Layout from "@/components/layout"
+import Layout from "@/components/layout/layout"
 import Head from "next/head"
 
 const App = ({ Component, pageProps }: AppProps) => {
