+++ styles/layout/ChatButton.module.scss
@@ -1,4 +1,4 @@
-@import './import';
+@import '../import';
 
 .chatBox{
     position: fixed;
@@ -9,4 +9,5 @@
     background-color: $primary-color;
     border: 1px solid $tertiary-color;
     z-index: 1;
+    display: none;
 }
