+++ styles/layout/Footer.module.scss
@@ -1,4 +1,4 @@
-@import './import';
+@import '../import';
 
 .footer{
     width: 100%;
