+++ styles/layout/Header.module.scss
@@ -1,4 +1,4 @@
-@import './import';
+@import '../import';
 
 .header{
     position: fixed;
