+++ styles/layout/Layout.module.scss
@@ -1,4 +1,4 @@
-@import './import';
+@import '../import';
 
 .main{
     color: $secondary-color;
