+++ styles/layout/Sidebar.module.scss
@@ -1,4 +1,4 @@
-@import './import';
+@import '../import';
 
 .navigation{
     position: fixed;
